+++ qwen3vl_node.py
@@ -30,9 +30,6 @@ class Qwen3VL_GGUF_Node:
                  "temperature": ("FLOAT", {"default": 0.7, "min": 0.0, "max": 2.0, "step": 0.01}),
                  "seed": ("INT", {"default": 42}),
                  "unload_all_models": ("BOOLEAN", {"default": False}),
-             },
-             "optional": {
-                 "user_prompt_prefix": ("STRING", {"default": "", "multiline": False, "forceInput": True}),
              }
          }
 
@@ -40,7 +37,7 @@ class Qwen3VL_GGUF_Node:
     FUNCTION = "run"
     CATEGORY = "multimodal/Qwen"
 
-    def run(self, image, system_prompt, user_prompt, model_path, mmproj_path, output_max_tokens, image_max_tokens, ctx, n_batch, gpu_layers, temperature, seed, unload_all_models, user_prompt_prefix=None):
+    def run(self, image, system_prompt, user_prompt, model_path, mmproj_path, output_max_tokens, image_max_tokens, ctx, n_batch, gpu_layers, temperature, seed, unload_all_models):
         
         if unload_all_models == True:
             comfy.model_management.unload_all_models()
@@ -67,10 +64,6 @@ class Qwen3VL_GGUF_Node:
         # Путь к скрипту (рядом с этим файлом)
         node_dir = os.path.dirname(os.path.abspath(__file__))
         script_path = os.path.join(node_dir, "qwen3vl_run.py")
-
-        if user_prompt_prefix != None:
-            if user_prompt_prefix != "":
-                user_prompt = user_prompt_prefix + " " + user_prompt
 
         # Создаём временный JSON-файл с параметрами
         config = {
@@ -144,17 +137,23 @@ class MasterPromptLoader:
     def INPUT_TYPES(cls):
         # Загружаем system prompts и user styles
         system_prompts = cls._load_section("_system_prompts")
-        user_styles = cls._load_section("_user_prompt_styles")
-
-        preset_names = list(system_prompts.keys())
+        system_names = list(system_prompts.keys())
         
+        user_styles = cls._load_section("_user_prompt_styles")
         style_names = ["No changes"] + list(user_styles.keys())
 
         return {
             "required": {
-                "master_preset": (preset_names, ),
+                "master_preset": (system_names, ),
                 "style_preset": (style_names, {"default": "No changes"}),
+                "caption_length": ([
+                    "any", "very_short", "short", "medium", "long", "very_long"
+                ] + [str(i) for i in range(20, 261, 10)], {"default": "long"}),
+            },
+            "optional": {
+                "custom_user_prompt": ("STRING", {"default": "", "multiline": True}),
             }
+
         }
 
     RETURN_TYPES = ("STRING","STRING")
@@ -164,29 +163,220 @@ class MasterPromptLoader:
 
     @staticmethod
     def _load_section(section_key):
-        """Загружает указанную секцию из system_prompts.json."""
         current_dir = os.path.dirname(os.path.abspath(__file__))
-        json_path = os.path.join(current_dir, "system_prompts.json")
+        official_path = os.path.join(current_dir, "system_prompts.json")
+        user_path = os.path.join(current_dir, "system_prompts_user.json")
 
-        if not os.path.exists(json_path):
-            raise FileNotFoundError(f"system_prompts.json not found at {json_path}")
+        official_data = {}
+        if os.path.exists(official_path):
+            with open(official_path, "r", encoding="utf-8") as f:
+                official_data = json.load(f)
+        else:
+            print(f"[WARNING] Official prompt file not found: {official_path}")
 
-        with open(json_path, "r", encoding="utf-8") as f:
-            data = json.load(f)
+        user_data = {}
+        if os.path.exists(user_path):
+            with open(user_path, "r", encoding="utf-8") as f:
+                user_data = json.load(f)
 
-        return data.get(section_key, {})
+        # Получаем секции
+        official_section = official_data.get(section_key, {})
+        user_section = user_section = user_data.get(section_key, {})
 
-    def load_prompt(self, master_preset, style_preset):
+        combined = {**official_section, **user_section}
+        return combined
+
+    def load_prompt(self, 
+        master_preset, 
+        style_preset,
+        caption_length,
+        custom_user_prompt=""):
+
         system_prompts = self._load_section("_system_prompts")
-        user_styles = self._load_section("_user_prompt_styles")
+        system_prompt = system_prompts.get(master_preset, "").strip()
 
-        base_prompt = system_prompts.get(master_preset, "").strip()
-        system_prompt = base_prompt
+        #-------
 
         user_prompt = ""
+
+        # length
+        if caption_length not in ["any", "very_short", "short", "medium", "long", "very_long"]:
+            user_prompt += f"Keep it within {caption_length} words."
+        elif caption_length != "any":
+            user_prompt += f"Make it a {caption_length.replace('_', '-')} caption."
+
+        # style
+        user_styles = self._load_section("_user_prompt_styles")
         if style_preset != "No changes":
             style_instruction = user_styles.get(style_preset, "").strip()
-            user_prompt = style_instruction
+            user_prompt += '\n' + style_instruction
+
+        # custom_user_prompt
+        if custom_user_prompt != None:
+            if custom_user_prompt.strip():
+                user_prompt += "\n" + custom_user_prompt.strip()
+
+        return (system_prompt,user_prompt)
+
+class MasterPromptLoaderAdvanced:
+    @classmethod
+    def INPUT_TYPES(cls):
+        # Загружаем system prompts и user styles
+        system_prompts = cls._load_section("_system_prompts")
+        system_names = list(system_prompts.keys())
+        
+        user_styles = cls._load_section("_user_prompt_styles")
+        style_names = ["No changes"] + list(user_styles.keys())
+
+        return {
+            "required": {
+                "master_preset": (system_names, ),
+                "style_preset": (style_names, {"default": "No changes"}),
+                "caption_length": ([
+                    "any", "very_short", "short", "medium", "long", "very_long"
+                ] + [str(i) for i in range(20, 261, 10)], {"default": "long"}),
+            },
+            "optional": {
+                "custom_user_prompt": ("STRING", {"default": "", "multiline": True}),
+                "character_name": ("STRING", {"default": "", "multiline": False, "tooltip": "The name to use for the character"}),
+                "describe_lighting": ("BOOLEAN", {"default": False, "tooltip": "Include details about lighting: natural/artificial, soft/harsh, direction, and mood."}),
+                "describe_camera_angle": ("BOOLEAN", {"default": False, "tooltip": "Specify the camera perspective: eye-level, low-angle, bird’s-eye view, etc."}),
+                "describe_depth_of_field": ("BOOLEAN", {"default": False, "tooltip": "Describe focus and blur: e.g., “shallow depth of field,” “background blurred,” or “everything in focus.”"}),
+                "describe_composition": ("BOOLEAN", {"default": False, "tooltip": "Analyze visual structure: rule of thirds, symmetry, leading lines, balance, framing."}),
+                "describe_facial_details": ("BOOLEAN", {"default": False, "tooltip": "Provide a detailed description of facial features (eyes, mouth, expression) and the emotional state of any characters."}),
+                "describe_artistic_style": ("BOOLEAN", {"default": False, "tooltip": "Clearly identify and describe the artistic or rendering style of the image (e.g., photorealistic, anime, oil painting, pixel art, 3D render)."}),
+                "rate_aesthetic_quality": ("BOOLEAN", {"default": False, "tooltip": "Add a subjective quality rating: e.g., “low quality,” “high quality,” or “masterpiece.”"}),
+                "detect_watermark": ("BOOLEAN", {"default": False, "tooltip": "State whether a visible watermark is present in the image."}),
+                "skip_fixed_traits": ("BOOLEAN", {"default": False, "tooltip": "Avoid mentioning unchangeable attributes like ethnicity, gender, or age. Promotes ethical and flexible descriptions."}),
+                "skip_resolution": ("BOOLEAN", {"default": False, "tooltip": "Do not mention image resolution (e.g., “4K,” “1080p”)."}),
+                "ignore_image_text": ("BOOLEAN", {"default": False, "tooltip": "Do not describe any visible text, logos, or captions in the image."}),
+                "use_precise_language": ("BOOLEAN", {"default": False, "tooltip": "Avoid vague terms like “something” or “kind of.” Use specific, concrete descriptions."}),
+                "family_friendly": ("BOOLEAN", {"default": False, "tooltip": "Keep the caption suitable for all audiences (PG/SFW). No sexual, violent, or mature content."}),
+                "classify_content_rating": ("BOOLEAN", {"default": False, "tooltip": "Explicitly label the image as sfw, suggestive, or nsfw."}),
+                "focus_on_key_elements": ("BOOLEAN", {"default": False, "tooltip": "Describe only the most important subjects — omit background clutter, minor details, or decorations."}),
+                "European_woman": ("BOOLEAN", {"default": False, "tooltip": "Only if a woman is visibly present in the image, refer to her as 'European woman'."}),
+            }
+
+        }
+
+    RETURN_TYPES = ("STRING","STRING")
+    RETURN_NAMES = ("system_prompt","user_prompt")
+    FUNCTION = "load_prompt"
+    CATEGORY = "multimodal/Qwen"
+
+    @staticmethod
+    def _load_section(section_key):
+        current_dir = os.path.dirname(os.path.abspath(__file__))
+        official_path = os.path.join(current_dir, "system_prompts.json")
+        user_path = os.path.join(current_dir, "system_prompts_user.json")
+
+        official_data = {}
+        if os.path.exists(official_path):
+            with open(official_path, "r", encoding="utf-8") as f:
+                official_data = json.load(f)
+        else:
+            print(f"[WARNING] Official prompt file not found: {official_path}")
+
+        user_data = {}
+        if os.path.exists(user_path):
+            with open(user_path, "r", encoding="utf-8") as f:
+                user_data = json.load(f)
+
+        # Получаем секции
+        official_section = official_data.get(section_key, {})
+        user_section = user_section = user_data.get(section_key, {})
+
+        combined = {**official_section, **user_section}
+        return combined
+
+
+    def load_prompt(self, 
+        master_preset, 
+        style_preset,
+        caption_length,
+        custom_user_prompt="",
+        character_name="",
+        describe_lighting=False,
+        describe_camera_angle=False,
+        describe_depth_of_field=False,
+        describe_composition=False,
+        describe_facial_details=False,
+        describe_artistic_style=False,
+        rate_aesthetic_quality=False,
+        detect_watermark=False,
+        skip_fixed_traits=False,
+        skip_resolution=False,
+        ignore_image_text=False,
+        use_precise_language=False,
+        family_friendly=False,
+        classify_content_rating=False,
+        focus_on_key_elements=False,
+        European_woman=False):
+
+        system_prompts = self._load_section("_system_prompts")
+        system_prompt = system_prompts.get(master_preset, "").strip()
+
+        #-------
+
+        user_prompt = ""
+
+        # length
+        if caption_length not in ["any", "very_short", "short", "medium", "long", "very_long"]:
+            user_prompt += f"Keep it within {caption_length} words."
+        elif caption_length != "any":
+            user_prompt += f"Make it a {caption_length.replace('_', '-')} caption."
+
+        # style
+        user_styles = self._load_section("_user_prompt_styles")
+        if style_preset != "No changes":
+            style_instruction = user_styles.get(style_preset, "").strip()
+            user_prompt += '\n' + style_instruction
+
+        # custom_user_prompt
+        if custom_user_prompt != None:
+            if custom_user_prompt.strip():
+                user_prompt += "\n" + custom_user_prompt.strip()
+
+        # instructions
+        instructions = []
+        if character_name != None:
+            if character_name.strip() != "":
+                instructions.append(f"If a person is present, refer to them as '{character_name.strip()}'.")
+        if describe_lighting:
+            instructions.append("Include details about the lighting.")
+        if describe_camera_angle:
+            instructions.append("Describe the camera angle.")
+        if describe_depth_of_field:
+            instructions.append("Specify the depth of field (e.g., background blurred or in focus).")
+        if describe_composition:
+            instructions.append("Comment on the composition style (e.g., rule of thirds, symmetry).")
+        if describe_facial_details:
+            instructions.append("Provide a detailed description of facial features (eyes, mouth, expression) and the emotional state of any characters.")
+        if describe_artistic_style:
+            instructions.append("Clearly identify and describe the artistic or rendering style of the image (e.g., photorealistic, anime, oil painting, pixel art, 3D render).")
+        if rate_aesthetic_quality:
+            instructions.append("Rate the aesthetic quality from low to very high.")
+        if detect_watermark:
+            instructions.append("State if there is a watermark.")
+        if skip_fixed_traits:
+            instructions.append("Focus on what people are doing or wearing, not who they appear to be.")
+        if skip_resolution:
+            instructions.append("Describe only the depicted scene, objects, and people — not the image quality, format, or technical attributes.")
+        if ignore_image_text:
+            instructions.append("Describe only visual elements such as objects, people, colors, and lighting. Completely ignore any text, logos, watermarks, or UI elements in the image.")
+        if use_precise_language:
+            instructions.append("Use precise and unambiguous language.")
+        if family_friendly:
+            instructions.append("Keep the description family-friendly (PG).")
+        if classify_content_rating:
+            instructions.append("Classify the image as 'sfw', 'suggestive', or 'nsfw'.")
+        if focus_on_key_elements:
+            instructions.append("Only describe the most important elements of the image.")
+        if European_woman:
+            instructions.append("Only if a woman is visibly present in the image, refer to her as 'European woman'. Do NOT mention or describe any woman if none is present. Never invent, assume, or add female figures.")
+
+        if instructions:
+            user_prompt += "\n" + "\n".join(instructions)
 
         return (system_prompt,user_prompt)