+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # nightly
-- experimental: add style preset
+- add advanced options
+- add style preset
 # V1.2
 - add system_prompt template
 # V1.1